Added an ability to call transmorgify directly and use a cropname
It is now possible to call transmorgify directly using URL parameters
key={secrethash}&path={path}[&cropname={cropname}]

key is the secret key generated using sha1(action_string + SECRET_KEY)

action_string is the "_r222" style action strings

path is what is usually requested using the normal 404 handler.

so if the URL is

    http://assets-dev.usatoday.com/media/life/picture_r222.jpg

to generate a 222 wide image out of picture.jpg

The path query parameter will be

    path=/media/life/picture_r222.jpg

The optional cropname parameter substitutes the action string with a
 -{cropname}.

So if the full URL is

   http://assets-dev.usatoday.com/transmogrify?key=d79a4619060ec8dc7a06c8cbb24468aa0b285c89&path=/media/life/picture_r222.jpg&cropname=medium

The following image will be created:

   http://assets-dev.usatoday.com/media/life/picture-medium.jpg

Calls to transmogrify directly will always generate a new image so it
is possible to regenerate an image this way if you bypass the CDN.  Granted our
CDN will continue serving the cached version of the image until it's cache
is busted.